add sell stock functionality
added sell ticker function to search component to send sell request to server and update after transaction and display updated data. made limit order input in the sell modal to be inactive when the checkbox is unchecked. added route function in the server code very similar to the buy route function.